Move existing code into research subdirectory
The goal of this package is to present a single traversal algorithm. Rather
than remove all of the existing work from the repository, I've moved it to
spiral/research to preserve it.
