+++ spiral/research/traverser.py
