+++ spiral/research/visitor/deque_visitor.py
